Add TCP options parsing and more
- Added tcp_options.c and tcp_options.h for parsing TCP options.
- Introduced tcp_segment struct to encapsulate TCP segment details.
- Updated TCP connection handling to utilize SYN cookies for half-open connections.
- Enhanced error handling and return codes for connection creation.
- Refactored TCP listener hash table to improve matching logic.
- Improved formatting targets in CMakeLists.txt for better code consistency.
- Added address_types.h for MAC and IPv4 address definitions.
- Adjusted various source files to incorporate new structures and functions.
- Updated timer functions to include seconds retrieval.
- Enhanced debug output for socket operations.

+++ src/hash.h
@@ -9,4 +9,4 @@
 
 void hash_init(void);
 uint32_t hash_table(const void *data, size_t len);
-uint32_t hash_syncookie(const void *data, size_t len, uint32_t time);
+uint32_t hash_syncookie(const void *data, size_t len, uint64_t time);
